changes default append value to False in add_formula and refactors tests
